Improved default exports
