+++ packages/wasi/src/bindings/browser.ts
@@ -4,7 +4,7 @@ import hrtime from "../polyfills/browser-hrtime";
 // @ts-ignore
 import path from "path-browserify";
 
-import { WASIBindings, WASIExitError, WASIKillError } from "../wasi";
+import { WASIBindings, WASIExitError, WASIKillError } from "../index";
 import getBigIntHrtime from "../polyfills/hrtime.bigint";
 
 const bindings: WASIBindings = {
